Fixed freezing problem
Logic in IF statement for placing food was wrong. Previously it wasn't
placing food anywhere on the X or Y axis where the snake was on the same
of either axis. This was causing a freeze when the snake spanned across
the entire board and it tried to place food.